Refactor and simplify Spreadsheet class
